I don't need no transpilation...
I don't need to syntax control.

Seriously, I realized that we're transpiling a ton of code, but not actually using any language features that need it. Node 4 supports everything we're using.

Transpiling makes the toolchain twice as complex as it needs to be, so bye-bye.

Part 1 of my sinister plan to remove gulp.